+++ src/extras/hairlines.js
@@ -468,4 +468,5 @@ return hairlines;
 })();
 
 /* loader wrapper */
+Dygraph._required('dygraphs/src/extras/hairlines.js', /* exports */ {});
 })();